♻️ [refactor] : 헤더 유저 드롭다운을 공통 컴포넌트로 교체
